feat: addition of search service
Search service will implement searching using Fuse JS library.
This will have the options and searchBase added to Fuse constructur.

The options are what we will use in Fuse to provide additional options
for searching.

SearchBase will have the geojson properties (name, and id) which will be
used to return the results.