feat: Component base class
Component base class all components will inherit from. This will add
createEvents and triggerEvent functions that can be used forevent
handling.

The constructor is used to add properties, events and placeholder attrs
to the component. On creation of a component, the createEvents instance
function will be triggered adding events to the component, ie. if events
exist.

The template will be added to the componentElement