feat: map component
addition of toggleLayer, toggleLayer function is used to toggle between
adding and removing layers that are selected from the map.

Addition of onEachKingdom function that is bound to each kingdom
Polygon, allows us to add click event handlers to each kingdom when they
are selected.

Addition of addKingdomGeoJson which will add a kingdom polygon to the
leaflet instance